+++ plugins/filaforge-huggingface-chat/src/Models/Setting.php
@@ -51,3 +51,4 @@ class Setting extends Model
 
 
 
+